+++ src/lib/components/Navbar.react.js
@@ -27,7 +27,7 @@ Navbar.propTypes = {
     /**
      *  Navbar content
      */
-    children: PropTypes.string,
+    children: PropTypes.node,
 
     /**
      *  Often used with CSS to style elements with common properties